Revert "pain"
This reverts commit 9a28b30134ec1de1fd1171a95c9678f3d383b9a3.

+++ Funky Monkey 4/src/trackingWheel.hpp
@@ -50,11 +50,11 @@ public:
     encoder = new pros::ADIEncoder(port, port+1, reverse);
   }
 
-  void setDistance(double idistance){
+  void setDistance(int idistance){
     distance = idistance;
   }
 
-  void setDiameter(double idiameter){
+  void setDiameter(int idiameter){
     diameter = idiameter;
   }
